+++ src/lib.rs
@@ -1,8 +1,6 @@
 #![doc = include_str!("../README.md")]
 
 mod append_vec;
-#[cfg(feature = "ABNF")]
-mod augmented;
 mod earley;
 mod error;
 mod expression;
@@ -18,7 +16,7 @@ pub use crate::production::Production;
 pub use crate::term::Term;
 
 #[cfg(feature = "ABNF")]
-pub use augmented::ABNF;
+pub use parsers::ABNF;
 pub use parsers::{Format, BNF};
 
 pub(crate) use hashbrown::HashMap;
