+++ src/parsers/mod.rs
@@ -1,3 +1,10 @@
+mod augmented;
+mod bnf;
+
+#[cfg(feature = "ABNF")]
+pub use augmented::ABNF;
+pub use bnf::BNF;
+
 use crate::expression::Expression;
 use crate::grammar::Grammar;
 use crate::production::Production;
@@ -5,7 +12,7 @@ use crate::term::Term;
 
 use nom::{
     branch::alt,
-    bytes::complete::{tag, take_till, take_until},
+    bytes::complete::{take_till, take_until},
     character::complete::{self, multispace0},
     combinator::{all_consuming, complete, eof, not, peek, recognize},
     error::VerboseError,
@@ -16,34 +23,7 @@ use nom::{
 
 pub trait Format {
     fn prod_lhs(input: &str) -> IResult<&str, Term, VerboseError<&str>>;
-
     fn nonterminal(input: &str) -> IResult<&str, Term, VerboseError<&str>>;
-}
-
-#[non_exhaustive]
-pub struct BNF;
-
-impl Format for BNF {
-    fn prod_lhs(input: &str) -> IResult<&str, Term, VerboseError<&str>> {
-        let (input, nt) =
-            delimited(complete::char('<'), take_until(">"), complete::char('>'))(input)?;
-
-        let (input, _) = preceded(complete::multispace0, tag("::="))(input)?;
-
-        Ok((input, Term::Nonterminal(nt.to_string())))
-    }
-
-    fn nonterminal(input: &str) -> IResult<&str, Term, VerboseError<&str>> {
-        let (input, nt) = complete(delimited(
-            complete::char('<'),
-            take_until(">"),
-            terminated(complete::char('>'), complete::multispace0),
-        ))(input)?;
-
-        not(complete(tag("::=")))(input)?;
-
-        Ok((input, Term::Nonterminal(nt.to_string())))
-    }
 }
 
 pub fn terminal(input: &str) -> IResult<&str, Term, VerboseError<&str>> {
